refactor(repository): Separate the category_id from ProductCreate schema
During the implementation of the tests for ProductRepository, it was
detected that forcing the `category_id` within the `ProductCreate` schema
was inflexible and complicated the tests.

This change modifies the signature of the `create_product` method so that it receives
the `category_id` as an explicit parameter.

This improves the separation of responsibilities.
- Now the 'ProductCreate' schema only contains product's exclusive data.
- Service layer will be responsible for obtainig, checking and passing the 'category_id' separately.